moved maintenance mode handling functions to misc
